+++ tests/OrderLimiterTest.php
@@ -29,12 +29,22 @@ class OrderLimiterTest extends TestCase {
 	/**
 	 * @test
 	 * @testdox get_interval() should return the interval setting
+	 * @group Intervals
 	 */
 	public function get_interval_should_return_the_interval_setting() {
 		update_option( OrderLimiter::OPTION_KEY, [
-			'interval' => 'daily',
+			'interval' => 'weekly',
 		] );
 
+		$this->assertSame( 'weekly', ( new OrderLimiter )->get_interval() );
+	}
+
+	/**
+	 * @test
+	 * @testdox get_interval() should default to "daily"
+	 * @group Intervals
+	 */
+	public function get_interval_should_default_to_daily() {
 		$this->assertSame( 'daily', ( new OrderLimiter )->get_interval() );
 	}
 
@@ -108,12 +118,15 @@ class OrderLimiterTest extends TestCase {
 	/**
 	 * @test
 	 * @testdox get_message() should replace the {current_interval} placeholder
+	 * @testWith ["{current_interval}"]
+	 *           ["{current_interval:date}"]
+	 * @group Placeholders
 	 */
-	public function get_message_should_replace_current_interval_placeholder() {
+	public function get_message_should_replace_current_interval_placeholder( $placeholder ) {
 		update_option( 'date_format', 'F j, Y' );
 		update_option( OrderLimiter::OPTION_KEY, [
 			'interval'        => 'weekly',
-			'customer_notice' => 'This started on {current_interval}',
+			'customer_notice' => "This started on {$placeholder}",
 		] );
 
 		$now     = new \DateTimeImmutable( 'now', wp_timezone() );
@@ -127,7 +140,29 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @testdox get_message() should replace the {current_interval:time} placeholder
+	 * @group Placeholders
+	 */
+	public function get_message_should_replace_current_interval_time_placeholder() {
+		update_option( 'time_format', 'g:ia' );
+		update_option( OrderLimiter::OPTION_KEY, [
+			'interval'        => 'hourly',
+			'customer_notice' => "This started at {current_interval:time}",
+		] );
+
+		$now     = new \DateTimeImmutable( 'now', wp_timezone() );
+		$limiter = new OrderLimiter( $now );
+
+		$this->assertSame(
+			'This started at ' . $limiter->get_interval_start()->format( 'g:ia' ),
+			$limiter->get_message( 'customer_notice' )
+		);
+	}
+
+	/**
+	 * @test
 	 * @testdox get_message() should replace the {limit} placeholder
+	 * @group Placeholders
 	 */
 	public function get_message_should_replace_limit_placeholder() {
 		update_option( OrderLimiter::OPTION_KEY, [
@@ -145,12 +180,15 @@ class OrderLimiterTest extends TestCase {
 	/**
 	 * @test
 	 * @testdox get_message() should replace the {next_interval} placeholder
+	 * @testWith ["{next_interval}"]
+	 *           ["{next_interval:date}"]
+	 * @group Placeholders
 	 */
-	public function get_message_should_replace_next_interval_placeholder() {
+	public function get_message_should_replace_next_interval_placeholder( $placeholder ) {
 		update_option( 'date_format', 'F j, Y' );
 		update_option( OrderLimiter::OPTION_KEY, [
 			'interval'        => 'weekly',
-			'customer_notice' => 'Check back on {next_interval}',
+			'customer_notice' => "Check back on {$placeholder}",
 		] );
 
 		$now     = new \DateTimeImmutable( 'now', wp_timezone() );
@@ -160,6 +198,87 @@ class OrderLimiterTest extends TestCase {
 			'Check back on ' . $limiter->get_next_interval_start()->format( 'F j, Y' ),
 			$limiter->get_message( 'customer_notice' )
 		);
+	}
+
+	/**
+	 * @test
+	 * @testdox get_message() should replace the {next_interval:time} placeholder
+	 * @group Placeholders
+	 */
+	public function get_message_should_replace_next_interval_time_placeholder() {
+		update_option( 'time_format', 'g:ia' );
+		update_option( OrderLimiter::OPTION_KEY, [
+			'interval'        => 'hourly',
+			'customer_notice' => "Check back at {next_interval:time}",
+		] );
+
+		$now     = new \DateTimeImmutable( 'now', wp_timezone() );
+		$limiter = new OrderLimiter( $now );
+
+		$this->assertSame(
+			'Check back at ' . $limiter->get_next_interval_start()->format( 'g:ia' ),
+			$limiter->get_message( 'customer_notice' )
+		);
+	}
+
+	/**
+	 * @test
+	 * @group Placeholders
+	 */
+	public function get_placeholders_should_return_an_array_of_default_placeholders() {
+		update_option( 'date_format', 'F j, Y' );
+		update_option( 'time_format', 'g:ia' );
+		update_option( OrderLimiter::OPTION_KEY, [
+			'interval' => 'hourly',
+		] );
+
+		$now          = new \DateTimeImmutable( '2020-04-27 12:15:00', wp_timezone() );
+		$current      = new \DateTimeImmutable( '2020-04-27 12:00:00', wp_timezone() );
+		$next         = new \DateTimeImmutable( '2020-04-27 13:00:00', wp_timezone() );
+		$placeholders = ( new OrderLimiter( $now ) )->get_placeholders();
+
+		$this->assertSame( $current->format( 'F j, Y' ), $placeholders['{current_interval}'] );
+		$this->assertSame( $current->format( 'F j, Y' ), $placeholders['{current_interval:date}'] );
+		$this->assertSame( $current->format( 'g:ia' ), $placeholders['{current_interval:time}'] );
+		$this->assertSame( $next->format( 'F j, Y' ), $placeholders['{next_interval}'] );
+		$this->assertSame( $next->format( 'F j, Y' ), $placeholders['{next_interval:date}'] );
+		$this->assertSame( $next->format( 'g:ia' ), $placeholders['{next_interval:time}'] );
+	}
+
+	/**
+	 * @test
+	 * @group Placeholders
+	 */
+	public function time_placeholders_should_replace_00_with_midnight() {
+		$this->markTestIncomplete( 'https://github.com/nexcess/limit-orders/issues/21' );
+
+		update_option( OrderLimiter::OPTION_KEY, [
+			'interval' => 'daily',
+		] );
+
+		$now          = new \DateTimeImmutable( '2020-04-27 12:15:00', wp_timezone() );
+		$current      = new \DateTimeImmutable( '2020-04-27 00:00:00', wp_timezone() );
+		$next         = new \DateTimeImmutable( '2020-04-28 00:00:00', wp_timezone() );
+		$placeholders = ( new OrderLimiter( $now ) )->get_placeholders();
+
+		$this->assertSame( __( 'midnight', 'limit-orders' ), $placeholders['{current_interval:time}'] );
+		$this->assertSame( __( 'midnight', 'limit-orders' ), $placeholders['{next_interval:time}'] );
+	}
+
+	/**
+	 * @test
+	 * @group Placeholders
+	 */
+	public function get_placeholders_should_filter_placeholders() {
+		add_filter( 'limit_orders_message_placeholders', function ( $placeholders ) {
+			$placeholders['{test}'] = 'Test value';
+
+			return $placeholders;
+		} );
+
+		$placeholders = ( new OrderLimiter() )->get_placeholders();
+
+		$this->assertSame( 'Test value', $placeholders['{test}'] );
 	}
 
 	/**
@@ -206,6 +325,27 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @group Intervals
+	 * @ticket https://github.com/nexcess/limit-orders/issues/18
+	 */
+	public function get_interval_start_for_hourly() {
+		update_option( OrderLimiter::OPTION_KEY, [
+			'interval' => 'hourly',
+		] );
+
+		$now   = new \DateTimeImmutable( '2020-04-27 12:05:00', wp_timezone() );
+		$start = new \DateTimeImmutable( '2020-04-27 12:00:00', wp_timezone() );
+
+		$this->assertSame(
+			$start->format( 'r' ),
+			( new OrderLimiter( $now ) )->get_interval_start()->format( 'r' ),
+			'Hourly intervals should start at the top of the hour.'
+		);
+	}
+
+	/**
+	 * @test
+	 * @group Intervals
 	 */
 	public function get_interval_start_for_daily() {
 		update_option( OrderLimiter::OPTION_KEY, [
@@ -224,6 +364,7 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @group Intervals
 	 */
 	public function get_interval_start_for_weekly() {
 		update_option( 'week_starts_on', 1 );
@@ -244,6 +385,7 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @group Intervals
 	 */
 	public function get_interval_start_for_weekly_with_a_non_standard_day() {
 		update_option( 'week_starts_on', 6 );
@@ -264,6 +406,7 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @group Intervals
 	 */
 	public function get_interval_start_for_weekly_when_today_is_the_first_day_of_the_week() {
 		update_option( 'week_starts_on', 1 );
@@ -284,6 +427,7 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @group Intervals
 	 */
 	public function get_interval_start_for_monthly() {
 		$today = new \DateTimeImmutable( 'now', wp_timezone() );
@@ -301,6 +445,7 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @group Intervals
 	 */
 	public function get_interval_start_should_be_idempotent() {
 		$now     = new \DateTimeImmutable( '00:00:00', wp_timezone() );
@@ -320,6 +465,27 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @group Intervals
+	 * @ticket https://github.com/nexcess/limit-orders/issues/18
+	 */
+	public function get_next_interval_start_for_hourly() {
+		update_option( OrderLimiter::OPTION_KEY, [
+			'interval' => 'hourly',
+		] );
+
+		$now  = new \DateTimeImmutable( '2020-04-27 12:05:00', wp_timezone() );
+		$next = new \DateTimeImmutable( '2020-04-27 13:00:00', wp_timezone() );
+
+		$this->assertSame(
+			$next->format( 'r' ),
+			( new OrderLimiter( $now ) )->get_next_interval_start()->format( 'r' ),
+			'The next hourly interval should begin at the top of the next hour.'
+		);
+	}
+
+	/**
+	 * @test
+	 * @group Intervals
 	 */
 	public function get_next_interval_start_for_daily() {
 		update_option( OrderLimiter::OPTION_KEY, [
@@ -339,6 +505,7 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @group Intervals
 	 */
 	public function get_next_interval_start_for_weekly() {
 		update_option( 'week_starts_on', 1 );
@@ -359,6 +526,7 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @group Intervals
 	 */
 	public function get_next_interval_start_for_monthly() {
 		update_option( OrderLimiter::OPTION_KEY, [
@@ -378,6 +546,27 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @group Intervals
+	 * @ticket https://github.com/nexcess/limit-orders/issues/18
+	 */
+	public function get_seconds_until_next_interval_for_hourly() {
+		update_option( OrderLimiter::OPTION_KEY, [
+			'interval' => 'hourly',
+		] );
+
+		$now  = new \DateTimeImmutable( '2020-04-27 12:05:00', wp_timezone() );
+		$next = new \DateTimeImmutable( '2020-04-27 13:00:00', wp_timezone() );
+
+		$this->assertSame(
+			$next->getTimestamp() - $now->getTimestamp(),
+			( new OrderLimiter( $now ) )->get_seconds_until_next_interval(),
+			'It should return the number of seconds until the next hour begins.'
+		);
+	}
+
+	/**
+	 * @test
+	 * @group Intervals
 	 */
 	public function get_seconds_until_next_interval_for_daily() {
 		update_option( OrderLimiter::OPTION_KEY, [
@@ -397,6 +586,7 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @group Intervals
 	 */
 	public function get_seconds_until_next_interval_for_weekly() {
 		update_option( 'week_starts_on', 1 );
@@ -417,6 +607,7 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @group Intervals
 	 */
 	public function get_seconds_until_next_interval_for_monthly() {
 		update_option( OrderLimiter::OPTION_KEY, [
@@ -641,8 +832,15 @@ class OrderLimiterTest extends TestCase {
 
 	/**
 	 * @test
+	 * @ticket https://github.com/nexcess/limit-orders/pull/13
 	 */
 	public function count_qualifying_orders_should_not_limit_results() {
+		update_option( OrderLimiter::OPTION_KEY, [
+			'enabled'  => true,
+			'interval' => 'daily',
+			'limit'    => 100,
+		] );
+
 		for ( $i = 0; $i < 24; $i++ ) {
 			$this->generate_order();
 		}