+++ CHANGELOG.md
@@ -3,6 +3,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+* Added "hourly" as a default interval for stores ([#20]).
+* Added new placeholders to user-facing messaging ([#20]):
+	- `{current_interval:date}` (alias of `{current_interval}`)
+	- `{current_interval:time}`
+	- `{next_interval:date}` (alias of `{next_interval}`)
+	- `{next_interval:time}`
+
+### Updated
+
+* The settings screen will now show custom placeholders that have been registered via the "limit_orders_message_placeholders" filter ([#20]).
+
 ## [Version 1.1.2] - 2020-04-17
 
 ### Fixed
@@ -44,3 +59,4 @@ Initial plugin release.
 [#8]: https://github.com/nexcess/limit-orders/pull/8
 [#10]: https://github.com/nexcess/limit-orders/pull/10
 [#13]: https://github.com/nexcess/limit-orders/pull/13
+[#20]: https://github.com/nexcess/limit-orders/pull/20